Add actions for tooltips and place input
When the user puts the stamp, a prompt appears that the request is being processed, when the request completes, it disappears. Also, if the request has failed, a hint appears, indicating this, and a field for entering the place manually. If the next request succeeds, the hint and the field disappears